fix auto upload

+++ flow/visualize/i210_replay.py
@@ -6,6 +6,7 @@ from copy import deepcopy
 import numpy as np
 import json
 import os
+import os.path
 import pytz
 import subprocess
 import time
@@ -241,7 +242,7 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
     metadata['submission_time'].append(cur_time)
     metadata['network'].append(network_name_translate(env.network.name.split('_20')[0]))
     metadata['is_baseline'].append(str(args.is_baseline))
-    if args.to_aws:
+    if args.use_s3:
         name, strategy = get_configuration(args.submitter_name, args.strategy_name)
         metadata['submitter_name'].append(name)
         metadata['strategy'].append(strategy)
@@ -362,8 +363,12 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
                 '{0}/test_time_rollout/{1}'.format(dir_path, emission_filename)
 
             output_path = os.path.join(output_dir, '{}-emission.csv'.format(exp_name))
-            # convert the emission file into a csv file
-            emission_to_csv(emission_path, output_path=output_path)
+            if os.path.exists(emission_path.replace('emission.xml', '0_emission.csv')):
+                # csv already exists
+                os.rename(emission_path.replace('emission.xml', '0_emission.csv'), output_path)
+            else:
+                # convert the emission file into a csv file
+                emission_to_csv(emission_path, output_path=output_path)
 
             # generate the trajectory output file
             trajectory_table_path = os.path.join(dir_path, '{}.csv'.format(source_id))
@@ -384,7 +389,8 @@ def replay(args, flow_params, output_dir=None, transfer_test=None, rllib_config=
             print("\nGenerated emission file at " + output_path)
 
             # delete the .xml version of the emission file
-            os.remove(emission_path)
+            if os.path.exists(emission_path):
+                os.remove(emission_path)
 
         all_trip_energies = os.path.join(output_dir, '{}-all_trip_energies.npy'.format(exp_name))
         np.save(all_trip_energies, dict(all_trip_energy_distribution))
@@ -510,6 +516,7 @@ def create_parser():
 
 
 def generate_graphs(args):
+    """Generate the graphs."""
     date = datetime.now(tz=pytz.utc)
     date = date.astimezone(pytz.timezone('US/Pacific')).strftime("%m-%d-%Y")
